% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plt.R
\name{label_thousands}
\alias{label_thousands}
\title{Thousands Labeller}
\usage{
label_thousands(x)
}
\arguments{
\item{x}{The number to be labeled}
}
\value{
A character string
}
\description{
Take a number, divide it by a thousand and add a capital "K"
}
