% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_05_2019.R
\name{plot_07_05_19}
\alias{plot_07_05_19}
\title{Create plot of student teacher ratios.}
\usage{
plot_07_05_19(animated = TRUE)
}
\arguments{
\item{animated}{Logical. Should an animated (gif) or static (png) plot be returned?}
}
\description{
Create a plot of GDP and student teacher ratios over time.
}
