% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_05_2019.R
\name{plot_15_05_19}
\alias{plot_15_05_19}
\title{Determine geographic variation in Nobel Prize winners}
\usage{
plot_15_05_19(start_year = 1969)
}
\arguments{
\item{start_year}{First year from which to start collating noble prize winners}
}
\value{
A ggplot plot grid
}
\description{
Determine geographic variation in Nobel Prize winners
}
