% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30_04_2019.R
\name{plot_30_4_19}
\alias{plot_30_4_19}
\title{Create plot on night time bird collisions}
\usage{
plot_30_4_19()
}
\value{
Produces a ggplot object, which is also saved as a .png in /plots
}
\description{
Task for 30/4/2019. Plot night time bird collisions in the US from
Winger et al. (2019) https://doi.org/10.1098/rspb.2019.0364
}
